Refactor common protocol code into NativeProtocol.swift
The new class called NativeProtocol now contains a protocol-agnostic implementation which is currently used by HTTPURLProtocol,
and in future will be used by other native protocols - FTP, Data etc.
